template/tools/task-board/requirements.txt
# Task Board Backend Dependencies
Flask==3.0.0
Flask-CORS==4.0.0
PyYAML==6.0.1
watchdog==3.0.0
